feat(hunt): Nearby moves under the map as a horizontal rail
Map and nearby caches now answer "what's around me" as one unit: the
Nearby section (header + search + filter) sits directly below the
mini-map and renders as a horizontal card rail (shared HuntRailCard,
extracted from HuntRecentlyAddedSection) instead of the old full-width
vertical list at the page bottom. Community rails follow. The shared
card brings the LP payout badge to both rails. Maestro testIDs from the
vertical-list era are preserved (hunt-discover-row-*, hunt-search-input,
hunt-filter-button, loading/empty states).
